Avoid holding all lighthouse data in memory
The script was starting to be killed because of memory usage.
As our lighthouse dataset has grown, holding it all in RAM is
unsustainable.

For now we're switching to holding all data for a particular
URL in RAM at a time. In future it might be worth being pickier
and only pulling out the data that's definitely needed.

This approach does lead to parsing each JSON file twice, so
could be a bit slower. It might be possible to remove that duplication
with a bit more exploration of the mapping between URLs in
list.txt and filenames for the lighthouse reports.